Parallelize CPU-intensive transforming step (#4)
Parallelism is configured by a new config parameter
`cpuParallelismFactor`. The actual parallelism is chosen dynamically
based on the number of available CPU, so the default value should be
appropriate for all sized VMs.

Also, `uploadConcurrency` is replaced with `uploadParallelismFactor`
to choose upload parallelism dynamically based on the number of
available CPU.