Recover from 'AlreadyExists' exception (#14)
We create unique file names for each data file.  But, if the loader
is recovering from an earlier network failure, then it might try
to upload the same file twice.

In DatabricksUploader, we set the option .setOverwrite(false), which
tells the SDK to raise this exception if the file already exists.

This is the correct design decision, because we don’t want to
double-trigger the LakeFlow (a.k.a. DLT) to load this file into
the Delta table.

But it means we need to catch and ignore the AlreadyExists exception.