Re-upload files after a AlreadyExists exception (#20)
The Databricks upload API sometimes writes corrupt files into the volume.

For example, upload attempt 1 results in a timeout or 5xx error
and upload attempt 2 results in a AlreadyExists exception.
In this scenario, we must not assume that upload attempt 1 was successful.
Because upload attempt 1 might have dumped a corrupt file in the volume.

This commit makes necessary changes to append retry count to the name of
the uploaded file. This makes filename unique in each re-upload.
We should never expect to get a AlreadyExists exception after this change.